migrated provinces

+++ db/migrate/20180403174924_create_provinces.rb
@@ -2,7 +2,7 @@ class CreateProvinces < ActiveRecord::Migration[5.1]
   def change
     create_table :provinces do |t|
       t.string :name
-      t.double :tax_rate
+      t.decimal :tax_rate
 
       t.timestamps
     end
